+++ src/database/migrations/2020_02_20_000624_create_purchases_items_table.php
@@ -15,8 +15,8 @@ class CreatePurchasesItemsTable extends Migration
     {
         Schema::create('purchases_items', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('purchases_id');
-            $table->integer('product_id');
+            $table->unsignedBigInteger('purchases_id');
+            $table->unsignedBigInteger('product_id');
             $table->integer('qty')->unsigned();
             $table->string('batch_number')->nullable();
             $table->timestamps();
