+++ src/database/migrations/2020_03_11_182830_create_orders_table.php
@@ -16,8 +16,8 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('name')->nullable();
-            $table->integer('client_id');
-            $table->integer('courier_id')->nullable();
+            $table->unsignedBigInteger('client_id');
+            $table->unsignedBigInteger('courier_id')->nullable();
             $table->string('tracking')->nullable();
             $table->integer('transaction_type_id');
             $table->date('date')->nullable();
